Update to v2.11
In the 'Report an Ad' feature, the extension will now perform a check for known malware
domains in the list of resources.  If resources from a known malware site are found, the
users is notified, and the a link with more information is provided.  If no malware
resources are found, the page continues to work as before. (#315)