Update to v2.12
Fixed an issue with the back button in Safari & YouTube. (#350)
Fixed an issue in the Chrome Beta version if the report an issue link in the popup menu. (#347)
In Chrome, improved the cleanup of old data when tabs are closed.(#345)
Fixed an issue with Whitelist wizard not being displayed on certain sites. (#341)
Fixed an issue with the AdBlock icon in Safari 5.0 (#342)
Added a work around to a Chrome bug for incorrectly reporting the type of a URL request. ((#339)
On the AdBlock options page, Filter Lists tab, add the title of custom filters lists the users is subscribed to (if there is a title for the filter list). (##336)
2.11: In the 'Report an Ad' feature, the extension will now perform a check for known malware
domains in the list of resources.  If resources from a known malware site are found, the
users is notified, and the a link with more information is provided.  If no malware
resources are found, the page continues to work as before. (#315)