Update to v2.13
Added the ability to exclude / disable filter list entries (#241)
Fixed an issue with a missing click handler on the popup menu (#359)
Remove confusing message on popup menu (#360)
Fixed an Javascript error on the background page in Safari when several tabs
are closed while still being renderred. (#357)
Added a user message that is displayed when the white list and black list
wizard can't run on older sites with framesets. (#352)
Added a 'reload page' option to the white list wizard. (#349)