Update to v2.16
AdBlock now uses the YouTube API to retreive the channel name when users have the Whitelist a YouTube Channel feature enabled. (#392)
The 'Filter syntax tutorial' link on the AdBlock options / Customize tab now checks if the generated URL is valid. (#394)
Added a check for chrome.runtime.onMessage API, since it isn't available in Chrome 25 and older versions. (#395)
Added an option to the General options tab to not display the ads block counter on the AdBlock popup menu.
The option to display the # of ads block was moved from the AdBlock popup menu to the General tab as well. (#401)
Added logic to set focus on new tabs open by AdBlock (for example, the options pages), when the Chrome is running in incognito mode. (#405)
On the 'Report an Ad' page, all of the questions are now bolded. (#416)
Reformated the text and fixed typos in the change log. (#418)
Modified several JavaScript throws statements to now use an Error object. (#420)