Update to v2.18
Fixed an issue when syncing custom filters using DropBox. (#413)
Fixed an issue with the number of blocked ads on single page sites. (#414)
Fixed an issue with the URL of Malware filter lists on the AdBlock Options page - Filter List tab. (#436)
Added an new Malware Notification option when users are subscribed to the Malware filter list. (#435)