Update to Firefox v2.4
Added the ping count to the ping data. (#159)
Removed the 'View the Resource' page. (#158)
Changed the order of IF statements in the 'shouldsurvey' function. (#157)
Fix a style issue with the 'Subscribe' popup. (#156)
Added more filters for showing ads on Google Search when that option is enabled. (#154)
Removed some Chrome specific code. (#152)