Update to v2.48
Updated Translations.
Updated the placehodler text in the "Show ads on web page or domain" example URL (#669).
Removed duplicate logic on the General tab of the AdBlock Options page (#671).
Migrated links to the new AdBlock help site (#632).
Updated the "Report an Ad" and "Bug Report" pages to use the new AdBlock help site (#632).