Update to v2.49
Updated Translations.
Removed old code that is no longer needed with migration to the new AdBlock help / support site (#684).
Fixed the AdBlock logo image on the "Show all requests" page (#678).
Added a check for the tab 'id' property in the AdBlock popup (#674).
Refactored the logic on the AdBlock Options page.  Also, for Safari users, added a message that Acceptable Ads and Content blocking are mutual exclusive (#682).