Update to v3.0.4
Fixed a bug on 'Options' page when the page was openned prior to the background page being initialized. (#112)
Fixed a bug when the AdBlock context menu option was enabled / disabled on the Options page.  (#111)
Fixed a bug on 'Report an Ad' page for non-English users reporting an ad on English site.  (#108)
Updated the descriptions for certain language specific filter lists to include 'EasyList'.  (#107)