Update to v1.3.0.0
Added support of blocking ads that are loaded using web sockets.  (#71)
Improved the security for the 'bandaids' processing on the getadblock.com site by improving the host name check, and adding a check to the click event handler for isTrusted.  (#170)
Updated the Indonesian filter list URL.  (#67)
Added support of the generic hide and generic block filter list options.  (#73)
Changed the localized description string in the locale files.  (#80)