Update to v3.4.0
Fixed an issue with the lower section of the AdBlock Options page being hidden behind the footer. (#167)
Improved the security for the 'bandaids' processing on the getadblock.com/installed page by improving the host name check, and adding a check to click event for isTrusted.  (#170)
Updated the 'stats' processing to store the next ping time, ping count, and unique identifier to two different persisted storage areas.  (#174)