Update to v2.3.0.0
Fixed an issue with downloading certain Filters Lists. (#158)
Updated translations
Changed all hard coded links in AdBlock from HTTP to HTTPS (SSL) (when possible).  (#163)
Added a warning message to the Filter Lists tab on the Options page when the user un-subscribes to EasyList and / or all filter lists. (#159)
Removed old 1.X to 2.X upgrade logic.  (#160)
Fixed an issue on the Customize tab of the AdBlock Options page with validation of user entered filter list rules.  (#167)
Added support for new filter list syntax. (#156)
Updated the AdBlock pop up menu to now have a link for Help and Feedback. (#162)
Modified the method of insert the unique identifier into getadblock.com pages. (#150)
Updated the success message on the "report an Ad" page. (#149)
Reduced the usage of Regexps to reduce the memory usage.  (#155)